Fix : 언어 정보 정렬 바꿈

+++ backend/src/main/java/com/ssafy/backend/domain/analysis/service/AnalysisGithubService.java
@@ -55,7 +55,7 @@ public class AnalysisGithubService {
 	private List<GithubDetailLanguage> getDetailLanguages(Github github) {
 		return githubLanguageQueryRepository.findByGithub(github)
 			.stream()
-			.sorted(Comparator.comparingInt(l -> -Integer.parseInt(l.getPercentage())))
+			.sorted(Comparator.comparingDouble(l -> -Double.parseDouble(l.getPercentage())))
 			.collect(Collectors.toList());
 
 	}
